Agrega formulario para agregar o editar un usuario
