+++ authorization/urls.py
@@ -12,13 +12,13 @@ from authorization.views.permissions import (
     PermissionListView, PermisoUpdateView
 )
 from authorization.views.users import (
-    UserListView, UserCreateView, UserDetailView, UserUpdateView, UserDeleteView
+    UserListView, UserCreateFormView, UserDetailView, UserUpdateView, UserDeleteView
 )
 
 urlpatterns = [
     # Usuarios
     path('usuario/', UserListView.as_view(), name='user-list'),
-    path('usuario/agregar', UserCreateView.as_view(), name='user-create'),
+    path('usuario/agregar', UserCreateFormView.as_view(), name='user-create'),
     path('usuario/<int:pk>/', UserDetailView.as_view(), name='user-detail'),
     path('usuario/<int:pk>/editar/', UserUpdateView.as_view(), name='user-update'),
     path('usuario/<int:pk>/eliminar/', UserDeleteView.as_view(), name='user-delete'),
